Move copyFile from diffUtil to MetadataFiles
Remove some static string
