+++ src/shared/metadataFiles.ts
@@ -9,6 +9,7 @@ import FileUtils from "./fileutils";
 import _ from "lodash";
 import ignore from "ignore";
 import * as fs from "fs";
+import * as glob from "glob";
 
 export default class MetadataFiles {
   public static sourceOnly: boolean = false;
@@ -148,5 +149,148 @@ export default class MetadataFiles {
   //Check if a component is accepted by forceignore.
   private accepts(filePath: string) {
     return !this.forceignore.ignores(path.relative(process.cwd(), filePath));
+  }
+
+  /**
+   * Copy a file to an outpu directory. If the filePath is a Metadata file Path,
+   * All the metadata requirement are also copied. For example MyApexClass.cls-meta.xml will also copy MyApexClass.cls.
+   * Enforcing the .forceignore to ignire file ignored in the project.
+   * @param filePath
+   * @param outputFolder
+   */
+  public static copyFile(filePath: string, outputFolder: string) {
+    const LWC_IGNORE_FILES = ["jsconfig.json", ".eslintrc.json"];
+    const pairStatResources = METADATA_INFO.StaticResource.directoryName;
+    const pairStatResourcesRegExp = new RegExp(pairStatResources);
+    const pairAuaraRegExp = new RegExp(
+      METADATA_INFO.AuraDefinitionBundle.directoryName
+    );
+
+    let copyOutputFolder = outputFolder;
+
+    let exists = fs.existsSync(path.join(outputFolder, filePath));
+    if (exists) {
+      return;
+    }
+
+    if (filePath.startsWith(".")) {
+      fs.copyFileSync(filePath, outputFolder);
+      return;
+    }
+
+    let fileName = path.parse(filePath).base;
+    //exclude lwc ignored files
+    if (LWC_IGNORE_FILES.includes(fileName)) {
+      return;
+    }
+
+    let filePathParts = filePath.split(path.sep);
+
+    if (fs.existsSync(outputFolder) == false) {
+      fs.mkdirSync(outputFolder);
+    }
+    // Create folder structure
+    for (let i = 0; i < filePathParts.length - 1; i++) {
+      let folder = filePathParts[i].replace('"', "");
+      outputFolder = path.join(outputFolder, folder);
+      if (fs.existsSync(outputFolder) == false) {
+        fs.mkdirSync(outputFolder);
+      }
+    }
+
+    // Copy all file with same base name
+    let associatedFilePattern = "";
+    if (SOURCE_EXTENSION_REGEX.test(filePath)) {
+      associatedFilePattern = filePath.replace(SOURCE_EXTENSION_REGEX, ".*");
+    } else {
+      let extension = path.parse(filePath).ext;
+      associatedFilePattern = filePath.replace(extension, ".*");
+    }
+    let files = glob.sync(associatedFilePattern);
+    for (let i = 0; i < files.length; i++) {
+      if (fs.lstatSync(files[i]).isDirectory() == false) {
+        let oneFilePath = path.join(".", files[i]);
+        let oneFilePathParts = oneFilePath.split(path.sep);
+        fileName = oneFilePathParts[oneFilePathParts.length - 1];
+        let outputPath = path.join(outputFolder, fileName);
+        fs.copyFileSync(files[i], outputPath);
+      }
+    }
+
+    // Hadle ObjectTranslations
+    // If a file fieldTranslation is copied, make sure the ObjectTranslation File is also copied
+    if (
+      filePath.endsWith("Translation-meta.xml") &&
+      filePath.indexOf("globalValueSet") < 0
+    ) {
+      let parentFolder = filePathParts[filePathParts.length - 2];
+      let objectTranslation =
+        parentFolder + METADATA_INFO.CustomObjectTranslation.sourceExtension;
+      let outputPath = path.join(outputFolder, objectTranslation);
+      let sourceFile = filePath.replace(fileName, objectTranslation);
+      if (fs.existsSync(sourceFile) == true) {
+        fs.copyFileSync(sourceFile, outputPath);
+      }
+    }
+
+    //FOR STATIC RESOURCES - WHERE THE CORRESPONDING DIRECTORY + THE ROOT META FILE HAS TO BE INCLUDED
+    if (pairStatResourcesRegExp.test(filePath)) {
+      outputFolder = path.join(".", copyOutputFolder);
+      let srcFolder = ".";
+      let staticRecourceRoot = "";
+      let resourceFile = "";
+      for (let i = 0; i < filePathParts.length; i++) {
+        outputFolder = path.join(outputFolder, filePathParts[i]);
+        srcFolder = path.join(srcFolder, filePathParts[i]);
+        if (filePathParts[i] === METADATA_INFO.StaticResource.directoryName) {
+          let fileOrDirname = filePathParts[i + 1];
+          let fileOrDirnameParts = fileOrDirname.split(".");
+          srcFolder = path.join(srcFolder, fileOrDirnameParts[0]);
+          outputFolder = path.join(outputFolder, fileOrDirnameParts[0]);
+          resourceFile =
+            srcFolder + METADATA_INFO.StaticResource.sourceExtension;
+          METADATA_INFO.StaticResource.sourceExtension;
+          staticRecourceRoot =
+            outputFolder + METADATA_INFO.StaticResource.sourceExtension;
+          if (fs.existsSync(srcFolder)) {
+            if (fs.existsSync(outputFolder) == false) {
+              fs.mkdirSync(outputFolder);
+            }
+          }
+          break;
+        }
+      }
+      if (fs.existsSync(srcFolder)) {
+        FileUtils.copyRecursiveSync(srcFolder, outputFolder);
+      }
+      if (fs.existsSync(resourceFile)) {
+        fs.copyFileSync(resourceFile, staticRecourceRoot);
+      }
+    }
+    //FOR AURA components and LWC components
+    if (pairAuaraRegExp.test(filePath)) {
+      outputFolder = path.join(".", copyOutputFolder);
+      let srcFolder = ".";
+      for (let i = 0; i < filePathParts.length; i++) {
+        outputFolder = path.join(outputFolder, filePathParts[i]);
+        srcFolder = path.join(srcFolder, filePathParts[i]);
+        if (filePathParts[i] === "aura" || filePathParts[i] === "lwc") {
+          let fileOrDirname = filePathParts[i + 1];
+          let fileOrDirnameParts = fileOrDirname.split(".");
+          srcFolder = path.join(srcFolder, fileOrDirnameParts[0]);
+          outputFolder = path.join(outputFolder, fileOrDirnameParts[0]);
+
+          if (fs.existsSync(srcFolder)) {
+            if (fs.existsSync(outputFolder) == false) {
+              fs.mkdirSync(outputFolder);
+            }
+          }
+          break;
+        }
+      }
+      if (fs.existsSync(srcFolder)) {
+        FileUtils.copyRecursiveSync(srcFolder, outputFolder);
+      }
+    }
   }
 }
